image and instance adaptor layer scaffolding
