+++ keras_rcnn/layers/__init__.py
@@ -1,6 +1,4 @@
-from .adaptors import (ImageAdaptor, InstanceAdaptor)
-
-from .losses import (RCNN, RPN)
+from .losses import (ImageAdaptor, InstanceAdaptor, RCNN, RPN)
 
 from .object_detection import (Anchor, ObjectProposal, ProposalTarget)
